requirements_streamlit.txt
# Core ML and Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep Learning
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
datasets>=2.0.0

# Image Processing
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-image>=0.18.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Web Application
streamlit>=1.25.0

# Utilities
tqdm>=4.62.0
requests>=2.25.0
huggingface-hub>=0.10.0

# Jupyter (for development)
ipywidgets>=7.6.0
jupyter>=1.0.0 